[JIT] Add selective backend lowering API
**Summary**
This commit adds a helper/utility to faciliate the selective lowering of
specific submodules within a module hierarchy to a JIT backend. The reason
that this is needed is that lowering a submodule of a scripted
module to a backend after the module has been scripted requires
adjusting its JIT type.

**Test Plan**
This commit refactors `NestedModuleTest` in `jit/test_backends.py` to
use this new selective lowering API.

**Fixes**
This commit fixes ##41432.

ghstack-source-id: 5a7f564adc4e5a2c08f5e34143d3a3de42ea825c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/43613